+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: gusgonza <gusgonza@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/09 13:35:36 by gusgonza          #+#    #+#             */
-/*   Updated: 2024/02/14 18:48:47 by gusgonza         ###   ########.fr       */
+/*   Updated: 2024/02/16 18:26:02 by gusgonza         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,11 +37,11 @@ char	*ft_strnstr(const char *str, const char *to_find, size_t len);
 char	*ft_strchr(const char *str, int c);
 char	*ft_strrchr(const char *str, int c);
 
-void	ft_bzero(void *str, int n);
+void	ft_bzero(void *str, size_t n);
 int		ft_memcmp(const void *s1, const void *s2, size_t n);
 void	*ft_memcpy(void *dest, const void *src, size_t n);
 void	*ft_memmove(void *dest, const void *src, size_t len);
-void	*ft_memchr(const void *str, int c, int n);
+void	*ft_memchr(const void *str, int c, size_t n);
 void	*ft_memset(void *dest, int c, size_t n);
 
 void	*ft_calloc(size_t nmemb, size_t size);
